cryptenroll: move load_volume_key_keyfile() to cryptenroll-password.c
Conceptually a keyfile and a password are pretty much the same thing,
hence put them in the same file.
