+++ src/cryptenroll/cryptenroll-password.h
@@ -4,5 +4,7 @@
 #include "cryptenroll.h"
 #include "shared-forward.h"
 
+int load_volume_key_keyfile(const EnrollContext *c, struct crypt_device *cd, struct iovec *ret_vk);
+
 int load_volume_key_password(const EnrollContext *c, struct crypt_device *cd, struct iovec *ret_vk);
 int enroll_password(const EnrollContext *c, struct crypt_device *cd, const struct iovec *volume_key);
